refactor using another approach without the need of having different ssr component
