+++ src/UI/extension/ExtensionLoginButton/ExtensionLoginButton.tsx
@@ -1,16 +1,24 @@
 import React from 'react';
-import globalStyles from 'assets/sass/main.scss';
-import styles from './extensionLoginButton.styles.scss';
+import { withStyles } from 'hocs/withStyles';
+import { isSSR } from 'utils/isSSR';
 import { ExtensionLoginButtonComponent } from './ExtensionLoginButtonComponent';
 
-export const ExtensionLoginButton: typeof ExtensionLoginButtonComponent = (
-  props
-) => {
-  return (
-    <ExtensionLoginButtonComponent
-      {...props}
-      globalStyles={globalStyles}
-      styles={styles}
-    />
-  );
-};
+export const ExtensionLoginButton = isSSR()
+  ? withStyles(ExtensionLoginButtonComponent, {
+      local: () =>
+        import(
+          'UI/extension/ExtensionLoginButton/extensionLoginButton.styles.scss'
+        )
+    })
+  : (props: typeof ExtensionLoginButtonComponent) => {
+      const globalStyles = require('assets/sass/main.scss');
+      const styles = require('./extensionLoginButton.styles.scss');
+
+      return (
+        <ExtensionLoginButtonComponent
+          {...props}
+          globalStyles={globalStyles}
+          styles={styles}
+        />
+      );
+    };
