+++ src/utils/isSSR.ts
@@ -1,0 +1,3 @@
+import { isWindowAvailable } from './isWindowAvailable';
+
+export const isSSR = () => !isWindowAvailable();
